modified price parameter to take floating point number

+++ models.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String
+from sqlalchemy import Column, Integer, String, Float
 from database import Base
 
 class Medicine(Base):
@@ -6,4 +6,4 @@ class Medicine(Base):
 
     id = Column(Integer, primary_key=True, index=True, autoincrement=True)
     name = Column(String, index=True)
-    price = Column(Integer)
+    price = Column(Float)
